+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/NameTags.java
@@ -3,9 +3,9 @@ package insane96mcp.iguanatweaksreborn.module.items;
 import insane96mcp.iguanatweaksreborn.InsaneSO;
 import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.util.ModNBTData;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
@@ -20,13 +20,15 @@ import net.minecraftforge.event.entity.living.LivingDeathEvent;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Name tags", description = "Retrieve name tags from name tagged entities.")
-@LoadFeature(module = Modules.Ids.ITEMS)
+@LoadFeature(module = Modules.Ids.ITEMS, description = "Retrieve name tags from name tagged entities.")
 public class NameTags extends Feature {
-    public static final TagKey<EntityType<?>> SHOULDNT_DROP_NAMETAG = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(InsaneSO.MOD_ID, "shouldnt_drop_nametag"));
+    public static final TagKey<EntityType<?>> SHOULDNT_DROP_NAMETAG = TagKey.create(Registries.ENTITY_TYPE, InsaneSO.location("shouldnt_drop_nametag"));
+
+    public static ResourceLocation HAS_NAME_TAG;
 
     public NameTags(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
+        HAS_NAME_TAG = this.createDataKey("has_name_tag");
     }
 
     @SubscribeEvent
@@ -39,12 +41,12 @@ public class NameTags extends Feature {
                 || target.isDeadOrDying())
             return;
 
-        if (target.getPersistentData().contains(InsaneSO.RESOURCE_PREFIX + "has_name_tag"))
+        if (ModNBTData.get(target, HAS_NAME_TAG, Boolean.class))
             dropNameTag(target.level(), target);
         if (event.getItemStack().hasCustomHoverName())
-            target.getPersistentData().putBoolean(InsaneSO.RESOURCE_PREFIX + "has_name_tag", true);
+            ModNBTData.put(target, HAS_NAME_TAG, true);
         else {
-            target.getPersistentData().remove(InsaneSO.RESOURCE_PREFIX + "has_name_tag");
+            ModNBTData.put(target, HAS_NAME_TAG, false);
             target.setCustomName(null);
         }
 
@@ -55,7 +57,7 @@ public class NameTags extends Feature {
         if (!this.isEnabled()
                 || event.getEntity().level().isClientSide
                 || event.getEntity().getType().is(SHOULDNT_DROP_NAMETAG)
-                || !event.getEntity().getPersistentData().contains(InsaneSO.RESOURCE_PREFIX + "has_name_tag"))
+                || !ModNBTData.get(event.getEntity(), HAS_NAME_TAG, Boolean.class))
             return;
 
         dropNameTag(event.getEntity().level(), event.getEntity());